Cleaned up the win percentage entry and replaced WinPercentage's field with Double

+++ src/jvmMain/kotlin/model/Matchup.kt
@@ -3,4 +3,4 @@ package model
 data class Matchup(val character: Character, val winPercentage: WinPercentage)
 
 @JvmInline
-value class WinPercentage(val percentage: Float)
+value class WinPercentage(val percentage: Double)
